src/gitlab/python/requirements.txt
mysql-connector-python==2.0.4
python-dateutil==2.8.2
requests==2.31.0
numpy==1.24.2